pyproject.toml
[tool.poetry]
name = "annotation-tools"
version = "0.1.0"
description = "Annotation web interface and experimentation/analysis scripts"
authors = [
    "Zachary Levonian <zach@digitalharbor.org>"
]
license = "MIT"
readme = "README.md"
packages = [{include = "experiment", from = "src"}]
repository = "https://github.com/levon003/annotation-tools.git"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
poetry = "1.6.1"
streamlit = "^1.23.1"
pandas = "^2.0.2"
scipy = "^1.11.0"
argon2-cffi = "^21.3.0"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
numpy = "^1.25.2"
spacy = "^3.6.1"
statsmodels = "^0.14.0"
tabulate = "^0.9.0"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
matplotlib = "^3.7.1"
black = "^22.12.0"
isort = "^5.12"
flake8 = "^6.0.0"
nbqa = "^1.6.0"
pre-commit = "^2.21.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
jupyterlab = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
      .eggs         # exclude a few common directories in the
    | .git          # root of the project
    | .github
    | .gitignore
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | venv
    | _build
    | buck-out
    | build
    | ci
    | data
    | dist
    | docs
    | docsrc
  )/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
virtual_env = "venv"
